requirements.txt
opencv-python
mediapipe==0.10.13
pyserial
numpy